+++ app/Http/Middleware/AuthorizeApiKey.php
@@ -1,0 +1,39 @@
+<?php
+namespace App\Http\Middleware;
+
+use Closure;
+use Ejarnutowski\LaravelApiKey\Models\ApiKey;
+use Ejarnutowski\LaravelApiKey\Models\ApiKeyAccessEvent;
+use Illuminate\Http\Request;
+use Ejarnutowski\LaravelApiKey\Http\Middleware\AuthorizeApiKey as AuthApiKey;
+
+class AuthorizeApiKey extends AuthApiKey
+{
+
+
+    /**
+     * Handle the incoming request
+     *
+     * @param Request $request
+     * @param Closure $next
+     * @return \Illuminate\Contracts\Routing\ResponseFactory|mixed|\Symfony\Component\HttpFoundation\Response
+     */
+    public function handle(Request $request, Closure $next)
+    {
+        $apiKey_para = $request->get('key');
+        $apiKey = ApiKey::getByKey($apiKey_para);
+
+        if ($apiKey instanceof ApiKey) {
+            $this->logAccessEvent($request, $apiKey);
+            return $next($request);
+        }
+
+        return response([
+            'errors' => [[
+                'message' => 'Unauthorized'
+            ]]
+        ], 401);
+    }
+
+
+}