+++ routes/api.php
@@ -28,12 +28,13 @@ Route::group(['middleware' => 'auth:api'], function () {
     Route::get('subscription/{user_id}/{offer_id}', 'Api\OfferController@subscription'); # get a specific subscription
     Route::get('user/email', 'Api\UserController@showFromEmail'); # get user by email
     Route::resource('user', 'Api\UserController');
-    Route::put('offer/ext/{institution}/{offer}', 'Api\OfferController@updateExternal');
+
 });
 
 // Additional routes that don't require authentication
 Route::get('offer', 'Api\OfferController@index');
 Route::get('offer/{offer}', 'Api\OfferController@show');
+Route::get('list/offer/short', 'Api\OfferController@indexForTiles');
 Route::get('offer/ext/{institution}/{offer}', 'Api\OfferController@showExternal');
 Route::get('institution', 'Api\InstitutionController@index');
 Route::get('institution/{institution}', 'Api\InstitutionController@show');
@@ -46,3 +47,7 @@ Route::get('bot/cookie/set', 'Api\CookieController@setCookie');
 
 // Filter-Tags
 Route::get('filter/tags', 'Api\FilterController@getTags');
+
+// Routes that require API Key Authentification
+Route::put('offer/ext/{institution}/{offer}', 'Api\OfferController@updateExternal')->middleware('auth.apikey');
+